requirements.txt
Beaker==1.6.4
Jinja2==2.7
MarkupSafe==0.23
bottle==0.12.8
bottle-cork==0.11.1
bottle-sqlalchemy==0.4.1
psycopg2==2.5.4
pycrypto==2.6.1
wsgiref==0.1.2
